To get the predicted number of total bytes from the sample, we need to un-bias it by following the logic in 
jeprof's `AdjustSamples`: https:github.com/jemalloc/jemalloc/blob/498f47e1ec83431426cdff256c23eceade41b4ef/bin/jeprof.in#L4064-L4074 
 
However, this algorithm is actually wrong: you actually need to unbias each sample _before_ you add them together, rather 
than adding them together first and then unbiasing the average allocation size. But the heap profile format in released versions of jemalloc 
does not give us access to each individual allocation, so this is the best we can do (and `jeprof` does the same). 
 
It usually seems to be at least close enough to being correct to be useful, but could be very wrong if for the same stack, there is a 
very large amount of variance in the amount of bytes allocated (e.g., if there is one allocation of 8 MB and 1,000,000 of 8 bytes) 
 
In the latest unreleased jemalloc sources from github, the issue is worked around by unbiasing the numbers for each sampled allocation, 
and then fudging them to maintain compatibility with jeprof's logic. So, once those are released and we start using them, 
this will become even more correct. 
 
For more details, see this doc: https:github.com/jemalloc/jemalloc/pull/1902 
 
And this gitter conversation between me (Brennan Vincent) and David Goldblatt: https:gitter.im/jemalloc/jemalloc?at=5f31b673811d3571b3bb9b6b

+++ src/prof_common/src/lib.rs
@@ -22,7 +22,7 @@ pub enum ProfStartTime {
 #[derive(Clone, Debug)]
 pub struct WeightedStack {
     pub addrs: Vec<usize>,
-    pub weight: usize,
+    pub weight: f64,
 }
 
 #[derive(Default)]
@@ -52,7 +52,7 @@ impl StackProfile {
 }
 pub struct SymbolTrieNode {
     pub name: String,
-    pub weight: usize,
+    pub weight: f64,
     links: Vec<usize>,
 }
 
@@ -64,7 +64,7 @@ impl WeightedSymbolTrie {
     fn new() -> Self {
         let root = SymbolTrieNode {
             name: "".to_string(),
-            weight: 0,
+            weight: 0.0,
             links: vec![],
         };
         let arena = vec![root];
@@ -102,7 +102,7 @@ impl WeightedSymbolTrie {
         }
         let next = SymbolTrieNode {
             name: next_name.to_string(),
-            weight: 0,
+            weight: 0.0,
             links: vec![],
         };
         let idx = self.arena.len();
